\name{aclp}
\alias{aclp}
\docType{data}
\title{Example data for btscs function}
\description{A subset of data from Alvarez et. al. (1996).}
\usage{data(aclp)}
\format{
  A data frame with 4126 observations on the following 7 variables.
  \describe{
    \item{\code{cname}}{Country name}
    \item{\code{country}}{Numeric country identifier}
    \item{\code{year}}{Year of observation}
	\item{\code{reg}}{A dichotomous variable coded 1 for dictatorship, 0 for democracy}
	\item{\code{gdpw}}{GDP/worker, 1985 prices}
	\item{\code{popg}}{Population growth}
	\item{\code{democ}}{A dichotomous variable coded 1 for democracy, 0 for dictatorship, (1-\code{reg})}
  }
}
\references{
Alvarez, M., J.A. Cheibub, F. Limongi and A. Przeworski. 1996. Classifying political regimes. Studies in Comparative International Development 31 (Summer): 1-37.
}
\keyword{datasets}
