% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{central}
\alias{central}
\title{Generate central values of a varaible}
\usage{
central(x, ...)
}
\arguments{
\item{x}{A vector of values.}

\item{...}{currently unimplemented.}

\item{n_unique_fac}{number of unique values below which the variable will be treated like a factor.}
}
\value{
A scalar value giving the central value of the variable.
}
\description{
Function generates central values of a variable.  If the variable is numeric with 
more than \code{n_unique_fac} values, then the median is returned.  If the variable is
a factor or numeric with fewer than or equal to \code{n_unique_fac} unique values, 
then the modal value is returned.
}
