% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{describe_data}
\alias{describe_data}
\title{Interactive List of Variables}
\usage{
describe_data(data)
}
\arguments{
\item{data}{A dataset to be described.}
}
\description{
Makes an interactive list of variables and their descriptive labels.  Uses the \pkg{DT} package to generate the table.
}
