% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_utility.R
\name{df.terms}
\alias{df.terms}
\title{Retrieve term degrees of freedom from model}
\usage{
df.terms(model, term, ...)
}
\arguments{
\item{model}{A model object}

\item{term}{A model term}

\item{...}{Other arguments passed down, currently not implemented}
}
\description{
Retrieve term degrees of freedom from model
}
