\name{FactorplotEx}
\alias{FactorplotEx}
\docType{data}
\title{Example Data for the factorplot function}
\description{Randomly generated data to showcase the functionality of the \code{factorplot} function}
\usage{data(FactorplotEx)}
\format{
  A data frame with 1000 observations on the following 3 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{x1}}{a factor with levels \code{first} \code{second} \code{third}}
    \item{\code{x2}}{a factor with levels \code{one} \code{two} \code{three} \code{four} \code{five}}
  }
}
\keyword{datasets}
