\name{france}
\alias{france}
\docType{data}
\title{Example data for factorplot function}
\description{A subset of data from the 1994 Eurobarometer for France}
\usage{data(france)}
\format{
  A data frame with 542 observations on the following 5 variables.
  \describe{
    \item{\code{lrself}}{respondent's left-right self-placement on a 1(left)-10(right) scale}
    \item{\code{male}}{a dummy variable coded 1 for males and 0 for females}
    \item{\code{age}}{respondent's age}
	\item{\code{vote}}{a factor indicating vote choice with levels PCF, PS, Green, RPR and UDF}
	\item{\code{retnat}}{a factor indicating the respondent's retrospective national economic evaluation with levels Better, Same and Worse}
	\item{\code{voteleft}}{a dichotomous variable where 1 indicates a vote for a left party, 0 otherwise}
  }
}
\references{
Reif, Karlheinz and Eric Marlier.  1997. \emph{Euro-barometer 42.0:  The First Year of the New European Union, November-December 1994}.   Inter-university Consortium for Political and Social Research (ICPSR) [distributor].
}
\keyword{datasets}
