% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{gen.ilikelihood}
\alias{gen.ilikelihood}
\title{Generated individual likelihood values for a GAMLSS Model}
\usage{
gen.ilikelihood(object, sum = TRUE)
}
\arguments{
\item{object}{An object of class \code{gamlss}.}

\item{sum}{Whether the summed or individual log-likelihood should be returned.}
}
\description{
This is a slight modification of the \code{gen-Likelihood} function 
from version 5.1-6 of the \pkg{gamlss} package.  The main modification
was to put a conditional flag that would permit the individual log-
likelihood values (rather than the summed log-likelihood) to be 
returned.
}
