% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{ggpie}
\alias{ggpie}
\title{Pie Charts with ggplot2}
\usage{
ggpie(data, variable, addPct = c("pie", "none", "legend"))
}
\arguments{
\item{data}{A data frame to pass to the ggplot function}

\item{variable}{A variable to be plotted.}

\item{addPct}{Where labels should be added - "none" gives no labels, "legend" addes percentages to the color legend, "pie" addes the legends to the pie pieces.}
}
\value{
a ggplot
}
\description{
This function produces pie charts with ggplot2.
}
