\name{InteractionEx}
\alias{InteractionEx}
\docType{data}
\title{Example Data for DAintfun}
\description{Data to execute example code for \code{DAintfun}}
\usage{data(InteractionEx)}
\format{
  A data frame with 500 observations on the following 4 variables.
  \describe{
    \item{\code{y}}{a numeric vector}
    \item{\code{x1}}{a numeric vector}
    \item{\code{x2}}{a numeric vector}
    \item{\code{z}}{a numeric vector}
  }
}
\details{
These are randomly generated data to highlight the functionality of \code{DAintfun}}
\keyword{datasets}
