% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{loessDeriv}
\alias{loessDeriv}
\title{Estimate Derivatives of LOESS Curve.}
\usage{
loessDeriv(obj, delta = 1e-05)
}
\arguments{
\item{obj}{An object of class \code{loess}.}

\item{delta}{Small change to be induced to estimate derivative.}
}
\value{
A vector of first derivative values evaluated at each original
x-value.
}
\description{
Estimates the first derivatives of the LOESS curve.
}
\author{
Dave Armstrong
}
