% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{makeHypSurv}
\alias{makeHypSurv}
\title{Make Hypothetical Predictions for Survey Data}
\usage{
makeHypSurv(l, obj, ...)
}
\arguments{
\item{l}{A named list where variable names are the names and values of the
variables are the values.  Combinations will be made with
\code{expand.grid}.}

\item{obj}{A model object estimated with \code{survreg}.}

\item{...}{currently not implemented.}
}
\value{
A data frame.
}
\description{
Calculates survival probabilities for hypothetical data.
}
\author{
Dave Armstrong
}
