% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{plot.alsos}
\alias{plot.alsos}
\title{Plotting method for ALSOS object}
\usage{
\method{plot}{alsos}(x, which_var = NULL, return_data = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{alsos}.}

\item{which_var}{The name of a raw variable that was scaled in the 
alsos procedure for which a meausrement function should be returned.}

\item{return_data}{Logical indicating whether the data should be returned
(if \code{TRUE}) or a plot generated (if \code{FALSE})}

\item{...}{arguments to be passed in, currently not implemented.}
}
\value{
A plot.
}
\description{
Makes a plot or optionally returns data for user-generated
plots from an alsos object
}
