% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{plot.secdiff}
\alias{plot.secdiff}
\title{Plotting Method for Second Difference Objects}
\usage{
\method{plot}{secdiff}(x, level = 0.95, ...)
}
\arguments{
\item{x}{An object of class \code{secdiff}}

\item{level}{The confidence level of the confidence interval(s)}

\item{...}{Other arguments to be passed down, currently not implemented.}
}
\description{
Plots the results of the \code{secondDiff} function.
}
