% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{powerTrans}
\alias{powerTrans}
\title{Power Transformation Function}
\usage{
powerTrans(x, transPower)
}
\arguments{
\item{x}{Vector of values to be transformed}

\item{transPower}{The power of the transformation}
}
\value{
A vector of transformed values
}
\description{
Power transformation function that treats everything with
absolute power transform < .01 as the log transform.
}
