% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_utility.R
\name{predictor.names}
\alias{predictor.names}
\title{Retrieve predictor names from model}
\usage{
predictor.names(model, ...)
}
\arguments{
\item{model}{A model object}

\item{...}{Other arguments passed down, currently not implemented}
}
\value{
A character vector of names of predictor variables from the model
}
\description{
Retrieve predictor names from model
}
