% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{print.glmc2}
\alias{print.glmc2}
\title{Print method for glmChange objects}
\usage{
\method{print}{glmc2}(x, ...)
}
\arguments{
\item{x}{Object of class \code{glmc2}}

\item{...}{Currently unimplemented.}
}
\description{
Print method for object of class \code{glmc2}.
}
