% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{print.iqq}
\alias{print.iqq}
\title{Print method for intQualQuant objects.}
\usage{
\method{print}{iqq}(x, ...)
}
\arguments{
\item{x}{Object of class \code{iqq}}

\item{...}{Currently unimplmeneted}
}
\description{
Print method for objects of class \code{iqq} calculated
with the \code{intQualQuant} function.
}
