% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_utility.R
\name{response}
\alias{response}
\title{Retrieve model response}
\usage{
response(model, ...)
}
\arguments{
\item{model}{A model object}

\item{...}{Other arguments passed down, currently not implemented}
}
\description{
Retrieve model response
}
