% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_utility.R
\name{responseName}
\alias{responseName}
\title{Retrieve response name from model}
\usage{
responseName(model, ...)
}
\arguments{
\item{model}{A model object}

\item{...}{Other arguments passed down, currently not implemented}
}
\value{
A vector of response variable name(s).
}
\description{
Retrieve response name from model
}
