% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{summary.balsos}
\alias{summary.balsos}
\title{Summry method for Bayesian ALSOS}
\usage{
\method{summary}{balsos}(object, ...)
}
\arguments{
\item{...}{Other arguments, currently unimplemented}

\item{x}{Object of class \code{balsos}}
}
\description{
summary method for objects of class \code{balsos}
}
