% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{summary.secdiff}
\alias{summary.secdiff}
\title{Summary for Second Difference Objects}
\usage{
\method{summary}{secdiff}(object, ..., level = 0.95, digits = 3)
}
\arguments{
\item{object}{An object of class \code{secdiff}}

\item{...}{Other arguments to be passed down to \code{summary}.}

\item{level}{Confidence level for the confidence intervals}

\item{digits}{Number of digits to print}
}
\description{
Summary method for objects of class \code{secdiff}.
}
