% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{sumStats}
\alias{sumStats}
\alias{sumStats.data.frame}
\alias{sumStats.survey.design}
\title{Summary Statistics}
\usage{
sumStats(data, vars, byvar = NULL, convertFactors = TRUE)
}
\arguments{
\item{data}{A data frame from which variables will be extracted.}

\item{vars}{A character vector of variable names.}

\item{byvar}{A character string giving a variable name of a stratifying variable.  The summaries of the \code{vars} will be provided for each level of \code{byvar}.}

\item{convertFactors}{Logical indicating whether factors should be converted to numeric first and then summarised.}
}
\value{
a matrix of summary statistics for variables, optionally by the values of another variable.

a vector of summary statistics for each variable or variable-group combination.
}
\description{
Provides summary statistics (mean, sd, quartiles, IQR, missing n, valid n)
for the variables in a data frame.
}
\examples{
data(aclp)
sumStats(aclp, vars=c("gdpw", "popg"), byvar="democ")

}
