% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_utility.R
\name{term.names}
\alias{term.names}
\title{Retrieve term names from model}
\usage{
term.names(model, ...)
}
\arguments{
\item{model}{A model object}

\item{...}{Other arguments passed down, currently not implemented}
}
\description{
Retrieve term names from model
}
