% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{tTest}
\alias{tTest}
\title{t-Test function}
\usage{
tTest(x, y, data, ...)
}
\arguments{
\item{x}{The dichotmous variable for the test.}

\item{y}{The interval/ratio variable for the test.}

\item{data}{A data frame where \code{x} and \code{y} can be found.}

\item{...}{Other arguments to be passed down to \code{t.test}.}
}
\value{
an object of class \code{tTest}
}
\description{
This function is a wrapper to the \code{t.test} function but produces more information in the output.
}
