% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{unformulate}
\alias{unformulate}
\title{Break Apart Model Formula}
\usage{
unformulate(form, keep_env = FALSE)
}
\arguments{
\item{form}{A formula}

\item{keep_inv}{Logical indicating whether the formula's
environment should be returned with the result}
}
\value{
A list with \code{termlabels} giving the rhs terms of the
model, \code{response} give the lhs of the model, \code{env} optionally 
giving the environment of the formula and \code{vars} a vector of the 
variable names implicated in the formula
}
\description{
Works as a sort of inverse to \code{reformulate} 
by breaking apart the formula into response and the term labels. 
It also returns the variable names of all of the variables 
implicated in the formula.
}
\details{
A sort of inverse of the \code{reformulate} function.
}
