% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc-package.R
\docType{data}
\name{aclp}
\alias{aclp}
\title{Example data for btscs function}
\format{
A data frame with 4126 observations on the following 7 variables.
\describe{ \item{cname}{Country name} \item{country}{Numeric
country identifier} \item{year}{Year of observation}
\item{reg}{A dichotomous variable coded 1 for dictatorship, 0 for
democracy} \item{gdpw}{GDP/worker, 1985 prices}
\item{popg}{Population growth} \item{democ}{A dichotomous
variable coded 1 for democracy, 0 for dictatorship, (1-\code{reg})} }
}
\description{
A subset of data from Alvarez et. al. (1996).
}
\references{
Alvarez, M., J.A. Cheibub, F. Limongi and A. Przeworski. 1996.
Classifying political regimes. Studies in Comparative International
Development 31 (Summer): 1-37.
}
\keyword{datasets}
