\encoding{UTF-8}
\name{DAMisc-package}
\alias{DAMisc-package}
\alias{DAMisc}
\docType{package}
\title{Dave Armstrong's Miscellaneous Functions}
\description{Functions to aid in the presentation of linear model results}
\details{
\tabular{ll}{
Package: \tab DAMisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-05-18\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
These are functions that help present linear model results.  Largely, the represent alternatives in presentation to other R packages.  For example, the \code{factorplot} function offers an alternative to David Firth's \code{qvcalc} package.  This function calculates and presents exact variances of all simple contrasts.  Both \code{DAintfun} and \code{DAintfun2} are alternative ways of presenting interactions between two continuous variables.  \code{DAintfun2} gives results in line with the suggestions in Brambor, Clark and Golder (2006). 
}
\author{Dave Armstrong (UW-Milwaukee, Department of Political Science)\cr
Maintainer: Dave Armstrong <davearmstrong.ps@gmail.com>
}
\references{
Brambor, T., W.R. Clark and M. Golder.  (2006) Understanding Interaction Models: Improving Empirical Analyses.  Political Analysis 14, 63-82.\cr
Berrym, W., M. Golder and D. Milton.  (2012) 
}
