% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc-package.R
\docType{data}
\name{france}
\alias{france}
\title{Example data for factorplot function}
\format{
A data frame with 542 observations on the following 5 variables.
\describe{ \item{lrself}{respondent's left-right self-placement on a
1(left)-10(right) scale} \item{male}{a dummy variable coded 1 for
males and 0 for females} \item{age}{respondent's age}
\item{vote}{a factor indicating vote choice with levels PCF, PS,
Green, RPR and UDF} \item{retnat}{a factor indicating the
respondent's retrospective national economic evaluation with levels Better,
Same and Worse} \item{voteleft}{a dichotomous variable where 1
indicates a vote for a left party, 0 otherwise} }
}
\description{
A subset of data from the 1994 Eurobarometer for France
}
\references{
Reif, Karlheinz and Eric Marlier.  1997. \emph{Euro-barometer
42.0: The First Year of the New European Union, November-December 1994}.
Inter-university Consortium for Political and Social Research (ICPSR)
[distributor].
}
\keyword{datasets}
