% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_utility.R
\name{has.intercept}
\alias{has.intercept}
\title{Identify if model has intercept}
\usage{
has.intercept(model, ...)
}
\arguments{
\item{model}{A model object}

\item{...}{Other arguments passed down, currently not implemented}
}
\value{
Logical value indicating whether the model has an intercept term
}
\description{
Identify if model has intercept
}
