% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damisc-package.R
\docType{data}
\name{InteractionEx}
\alias{InteractionEx}
\title{Example Data for DAintfun}
\format{
A data frame with 500 observations on the following 4 variables.
\describe{ \item{y}{a numeric vector} \item{x1}{a numeric
vector} \item{x2}{a numeric vector} \item{z}{a numeric
vector} }
}
\description{
Data to execute example code for \code{DAintfun}
}
\details{
These are randomly generated data to highlight the functionality of
\code{DAintfun}
}
\keyword{datasets}
