% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fox_utility.R
\name{is.aliased}
\alias{is.aliased}
\title{Is coefficient aliased}
\usage{
is.aliased(model)
}
\arguments{
\item{model}{A model object}
}
\description{
Identifies whether there are aliased coefficients in model
}
