% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{make_assoc_stats}
\alias{make_assoc_stats}
\alias{concordant}
\alias{discordant}
\alias{ord.gamma}
\alias{ord.somers.d}
\alias{tau.b}
\alias{lambda}
\alias{phi}
\alias{V}
\alias{simtable}
\alias{simrho}
\alias{simtab}
\title{Make Categorical Association Statistics}
\usage{
make_assoc_stats(x, y, ordinal = FALSE, n = 1000, weight = NULL)
}
\arguments{
\item{x}{The row-variable in a contingency table}

\item{y}{The column-variable in a contingency table}

\item{ordinal}{Logical indicating whether ordinal measures should also be returned.}

\item{n}{Number of iterations in the simulation.}

\item{weight}{Vector of weights used to generate the table.}
}
\value{
A matrix of statistics and p-values.
}
\description{
Makes several common measures of association for contingency tables.  The 
p-values are obtained through simulation.
}
\examples{
data(france)
make_assoc_stats(france$retnat, france$vote)

}
