% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{makeHypSurv}
\alias{makeHypSurv}
\title{Make Hypothetical Predictions for Survey Data}
\usage{
makeHypSurv(l, obj, ...)
}
\arguments{
\item{l}{A named list where variable names are the names and values of the
variables are the values.  Combinations will be made with
\code{expand.grid}.}

\item{obj}{A model object estimated with \code{survreg}.}

\item{...}{currently not implemented.}
}
\value{
A data frame with the variables in `l` and: 
\item{p.fail}{The probability of vailure given `time`}
\item{time}{The time at which the failure probability is evaluated. }
}
\description{
Calculates survival probabilities for hypothetical data.
}
\examples{
data(cabinet)
library(survival)
mod <- survreg(Surv(durat, censor) ~ invest + polar, data=cabinet, 
               dist="exponential")
l= list(invest = 0, 
        polar = c(5,15,25))
out <- makeHypSurv(l, mod)
head(out)

}
\author{
Dave Armstrong
}
