% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{panel.ci}
\alias{panel.ci}
\title{Lattice panel function for confidence intervals}
\usage{
panel.ci(x, y, subscripts, lower, upper, zl)
}
\arguments{
\item{x, y}{Data from the call to \code{xyplot}.}

\item{subscripts}{Variable used to created the juxtaposed panels.}

\item{lower, upper}{95\% lower and upper bounds of \code{y}.}

\item{zl}{Logical indicating whether or not a horizontal dotted line at zero
is desired.}
}
\description{
This panel function is defined to plot confidence intervals in a multi-panel
lattice display. Note, both lower and upper must be passed directly to
\code{xyplot} as they will be passed down to the prepanel function.
}
\author{
Dave Armstrong
}
