% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{plot.secdiff}
\alias{plot.secdiff}
\title{Plotting Method for Second Difference Objects}
\usage{
\method{plot}{secdiff}(x, level = 0.95, ...)
}
\arguments{
\item{x}{An object of class \code{secdiff}}

\item{level}{The confidence level of the confidence interval(s)}

\item{...}{Other arguments to be passed down, currently not implemented.}
}
\value{
A plot of the second difference on the x-axis and the index 
(i.e., observation number) on the y-axis. The plot has individual second
differences and 95% confidence intervals.
}
\description{
Plots the results of the \code{secondDiff} function.
}
\examples{
data(aclp)
mod <- glm(democ ~ gdpw + popg,
           data = subset(aclp, year == 1980), 
           family=binomial)

secdiff <- secondDiff(mod, c("gdpw", "popg"), data = subset(aclp, year == 1980))
(plot(secdiff))
}
