% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{poisfit}
\alias{poisfit}
\title{Scalar Measures of Fit for Poisson GLMs Models}
\usage{
poisfit(obj)
}
\arguments{
\item{obj}{A model of class \code{glm} with \code{family=poisson}.}
}
\value{
A named vector of scalar measures of fit
}
\description{
Calculates scalar measures of fit for models with count dependent variables
along the lines described in Long (1997) and Long and Freese (2005).
}
\details{
\code{poisfit} calculates scalar measures of fit (many of which are
pseudo-R-squared measures) to describe how well a model fits data with a
count dependent variable.
}
\references{
Long, J.S.  1997.  Regression Models for Categorical and Limited
Dependent Variables.  Thousand Oaks, CA: Sage.

Long, J.S. and J. Freese.  2005.  Regression Models for Categorical Outcomes
Using Stata, 2nd ed.  College Station, TX: Stata Press.
}
\author{
Dave Armstrong
}
