% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{print.ordChange}
\alias{print.ordChange}
\title{Print method for ordChange objects}
\usage{
\method{print}{ordChange}(x, ..., digits = 3)
}
\arguments{
\item{x}{Object of class \code{ordChange}}

\item{...}{Other arguments to be passed down to the function}

\item{digits}{Number of digits to print}
}
\description{
Print methods for objects of class \code{ordChange}
}
