% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{print.pre}
\alias{print.pre}
\title{Print method for objects of class pre}
\usage{
\method{print}{pre}(x, ..., sim.ci = 0.95)
}
\arguments{
\item{x}{An object of class \code{pre}.}

\item{...}{Other arguments passed to print, currently not implemented}

\item{sim.ci}{Coverage for the simulated confidence interval, if
\code{sim=TRUE} in the call to \code{pre}.}
}
\description{
Prints the output from an object of class \code{pre}.  The function prints
all components of the calculation and optionally simulated confidence
bounds.
}
\seealso{
\code{pre}
}
\author{
Dave Armstrong
}
