% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{probgroup}
\alias{probgroup}
\alias{probgroup.polr}
\alias{probgroup.multinom}
\title{Plog Probabilities by Group}
\usage{
probgroup(obj, ...)
}
\arguments{
\item{obj}{Object of class \code{polr} or \code{multinom} where appropraite.}

\item{...}{Currently not implemented.}
}
\value{
A plot giving the probability with which each observations is 
classified as its observed value.
}
\description{
Plots predicted probabilities by value of the dependent variable for
proportional odds logistic regression and multinomial logistic regression
models
}
\details{
Plots the predicted probabilities by value of the dependent variable.  Each
panel only includes the observations that had the identified value of the
dependent variable.
}
\author{
Dave Armstrong
}
