\name{probgroup}
\alias{probgroup}
\alias{probgroup.polr}
\alias{probgroup.multinom}
\title{Plog Probabilities by Group}
\description{Plots predicted probabilities by value of the dependent variable for proportional odds logistic regression and multinomial logistic regression models}
\usage{
}
\arguments{
}
\details{
}
\value{
}
\references{
}
\author{Dave Armstrong}
\examples{
}
