% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{test.balsos}
\alias{test.balsos}
\title{Testing Measurement Level Assumptions.}
\usage{
test.balsos(obj, cor.type = c("pearson", "spearman"))
}
\arguments{
\item{obj}{An object of class \code{balsos}.}

\item{cor.type}{Type of correlation to be used in the p-value calculation.}
}
\value{
Printed output giving the Bayesian p-value evaluating the null that
ther eis no interesting difference between the original values and the
optimally scaled values.
}
\description{
Uses the algorithm discussed in Armstrong and Jacoby 2018) to test the
intervality of a variable scaled by the Bayesian ALSOS method.
}
\author{
Dave Armstrong
}
