% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{tscslag}
\alias{tscslag}
\title{Lag a time-series cross-sectional variables}
\usage{
tscslag(dat, x, id, time, lagLength = 1)
}
\arguments{
\item{dat}{A data frame.}

\item{x}{A string identifying variable to be lagged.}

\item{id}{A string identifying the name of the cross-sectional identifier.}

\item{time}{A string identifying the name of the time variable.}

\item{lagLength}{The length of the lag, use negative values for leading
variables.}
}
\value{
A vector giving the lagged values of \code{x}.
}
\description{
Lags (or leads) a variable in a time-series corss-sectional dataset.
}
\author{
Dave Armstrong
}
