% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{xt}
\alias{xt}
\alias{xt.data.frame}
\alias{xt.survey.design}
\alias{print.xt}
\title{Cross-Tabulation of Weighted or Unweighted Data}
\usage{
xt(data, var, byvar = NULL)
}
\arguments{
\item{data}{Either a data frame or a survey design object.}

\item{var}{Row variable for the cross-tabular.}

\item{byvar}{Optional column variable for the cross-tabulation.  If \code{NULL}, a frequency and relative frequency distribution of \code{var} will be produced. 

Produces a cross-tabulation and Chi-square statistic for weighted or unweighted data.}
}
\value{
A list with two elements - table of class \code{tabyl} and the returned results from \code{svychisq}.
}
\description{
Cross-Tabulation of Weighted or Unweighted Data
}
