% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMisc_functions.R
\name{xt}
\alias{xt}
\alias{xt.data.frame}
\alias{xt.survey.design}
\alias{print.xt}
\title{Cross-Tabulation of Weighted or Unweighted Data}
\usage{
xt(data, var, byvar = NULL, controlvar = NULL, weight = NULL, ...)
}
\arguments{
\item{data}{Either a data frame or a survey design object.}

\item{var}{Row variable for the cross-tabular.}

\item{byvar}{Optional column variable for the cross-tabulation.  If \code{NULL}, a frequency and relative frequency distribution of \code{var} will be produced.}

\item{controlvar}{The name of a categorical control variable.}

\item{weight}{If using a data frame (rather than a survey design object), specifying the name of a weighting variable will for the function to create a survey design with probability weights equal to the weight variable and then use the survey design object to make the cross-tabulation.}

\item{...}{Other arguments to be passed down to \code{make_assoc_stats}.  You can use this to calculate different statistics.  By default, you get Chi-squared, Cramer's V and lambda.  With \code{ordinal = TRUE} you will also get Gamma, Tau-b and Somer's D. 

Produces a cross-tabulation and Chi-square statistic for weighted or unweighted data.}
}
\value{
A list with two elements - table of class \code{tabyl} and the returned results from \code{svychisq}.
}
\description{
Cross-Tabulation of Weighted or Unweighted Data
}
\examples{
data(france)
xt(france, "vote", "retnat")

}
